Group location methods together
